+++ src/main/java/in/oneton/idea/spring/assistant/plugin/suggestion/component/MavenReIndexingDependencyChangeSubscriber.java
@@ -7,72 +7,56 @@ import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleManager;
 import com.intellij.openapi.project.DumbService;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.project.ProjectManagerListener;
-import com.intellij.openapi.util.Disposer;
-import com.intellij.util.messages.MessageBusConnection;
+import com.intellij.openapi.startup.StartupActivity;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.idea.maven.project.MavenImportListener;
 
-import java.util.List;
-import java.util.Objects;
-
 import static in.oneton.idea.spring.assistant.plugin.misc.GenericUtil.moduleNamesAsStrCommaDelimited;
 
-public class MavenReIndexingDependencyChangeSubscriber implements ProjectManagerListener {
+public class MavenReIndexingDependencyChangeSubscriber implements StartupActivity {
 
     private static final Logger log = Logger.getInstance(MavenReIndexingDependencyChangeSubscriber.class);
 
-    private MessageBusConnection connection;
-
     @Override
-    public void projectOpened(@NotNull final Project project) {
-        // This will trigger indexing
-        final var service = SuggestionIndexerProjectService.getInstance(project);
+    public void runActivity(@NotNull final Project project) {
         LogUtil.debug(() -> log.debug("Subscribing to maven dependency updates for project " + project.getName()));
 
-        try {
-            this.connection = project.getMessageBus().connect();
-            this.connection.subscribe(MavenImportListener.TOPIC, this.handler(project, service));
-        } catch (final Throwable e) { //NOSONAR
-            log.error("Failed to subscribe to maven dependency updates for project " + project.getName(), e);
-        }
-    }
+        MavenImportListener mavenImportListener = (importedProjects, newModules) -> {
 
-    @NotNull
-    private MavenImportListener handler(final @NotNull Project project, final SuggestionIndexerProjectService service) {
-        return (importedProjects, newModules) -> {
             final boolean proceed = importedProjects.stream()
-                    .filter(proj -> project.getName().equals(proj.getName()))
-                    .anyMatch(proj -> proj.getDirectory().equals(project.getBasePath()));
+                    .filter(p -> project.getName().equals(p.getDisplayName()))
+                    .anyMatch(p -> p.getDirectory().equals(project.getBasePath()));
 
             if (proceed) {
                 LogUtil.debug(() -> log.debug("Maven dependencies are updated for project " + project.getName()));
-                DumbService.getInstance(project)
-                        .smartInvokeLater(() -> this.reindexModules(service, project, newModules));
+
+                DumbService.getInstance(project).smartInvokeLater(() -> {
+                    log.debug("Will attempt to trigger indexing for project " + project.getName());
+
+                    try {
+                        final Module[] modules = ModuleManager.getInstance(project).getModules();
+                        if (modules.length > 0) {
+                            SuggestionIndexerProjectService.getInstance(project).index(modules);
+                            return;
+                        }
+
+                        LogUtil.debug(() -> log.debug("Skipping indexing for project " + project.getName() + " as there are no modules"));
+                    } catch (final Throwable e) {
+                        log.error("Error occurred while indexing project " + project.getName() + " & modules " +
+                                moduleNamesAsStrCommaDelimited(newModules, false), e);
+                    }
+                });
+
+            } else {
+                log.debug("Skipping indexing as none of the imported projects match our project " + project.getName());
             }
         };
-    }
 
-    private void reindexModules(final SuggestionIndexerProjectService service, @NotNull final Project project, final @NotNull List<Module> newModules) {
-        log.debug("Will attempt to trigger indexing for project " + project.getName());
         try {
-            final Module[] modules = ModuleManager.getInstance(project).getModules();
-
-            if (modules.length > 0) {
-                service.index(modules);
-            }
-        } catch (final Throwable e) {//NOSONAR
-            log.error("Error occurred while indexing project " + project.getName() +
-                    " & modules " + moduleNamesAsStrCommaDelimited(newModules, false), e);
+            project.getMessageBus().connect().subscribe(MavenImportListener.TOPIC, mavenImportListener);
+        } catch (final Throwable e) {
+            log.error("Failed to subscribe to maven dependency updates for project " + project.getName(), e);
         }
-    }
-
-    @Override
-    public void projectClosed(@NotNull final Project project) {
-        if (Objects.nonNull(this.connection)) {
-            this.connection.disconnect();
-        }
-        Disposer.dispose(project);
     }
 
 }